+++ packages/ui/src/form/primitives/index.ts
@@ -11,5 +11,6 @@ export { Radio } from './radio';
 export * as Select from './select';
 export { SearchableChipSelect } from './searchable-chip-select';
 export { SearchableSelect } from './searchable-select';
+export { Switch } from './switch';
 export { Textarea } from './textarea';
 export { ValidityIndicator } from './validity-indicator';